+++ src/consts.js
@@ -11,10 +11,17 @@ exports.MOUSE_STEPS = 5;
 // 'document', 'image', 'xhr', 'script', 'stylesheet', 'font', 'other', 'manifest'
 exports.MEDIA_TYPES = ['image'];
 
+exports.LABELS = {
+    DETAIL: 'DETAIL',
+    MASTER: 'MASTER',
+    CHANNEL: 'CHANNEL',
+    SEARCH: 'SEARCH',
+};
+
 exports.SELECTORS = {
     SEARCH: {
         searchBox: 'input#search',
-        toggleFilterMenu: 'ytd-toggle-button-renderer a #button',
+        toggleFilterMenu: 'paper-button[aria-label="Search filters"]',
         filterBtnsXp: '//ytd-search-filter-renderer/a/div/yt-formatted-string',
         youtubeVideosSection: 'ytd-item-section-renderer',
         youtubeVideosRenderer: 'ytd-video-renderer,ytd-grid-video-renderer', // grid is for channels
@@ -22,7 +29,7 @@ exports.SELECTORS = {
     },
     VIDEO: {
         titleXp: '//ytd-video-primary-info-renderer/div/h1/yt-formatted-string',
-        viewCountXp: '//yt-view-count-renderer/span[1]',
+        viewCountXp: '//*[@id="count"]/ytd-video-view-count-renderer/span[1]',
         uploadDateXp: '//ytd-video-primary-info-renderer/div/div/div[1]/div[2]/yt-formatted-string',
         likesXp: "//ytd-menu-renderer/div/ytd-toggle-button-renderer[1]/a/*[@id='text']",
         dislikesXp: "//ytd-menu-renderer/div/ytd-toggle-button-renderer[2]/a/*[@id='text']",