+++ src/main.js
@@ -3,9 +3,12 @@ const Apify = require('apify');
 const utils = require('./utility');
 const crawler = require('./crawler_utils');
 
-const { log } = Apify.utils;
+const { log, puppeteer } = Apify.utils;
 
 Apify.main(async () => {
+    /**
+     * @type {any}
+     */
     const input = await Apify.getInput();
     // const input = {
     //   maxResults: 1,
@@ -18,59 +21,25 @@ Apify.main(async () => {
     //   }
     // };
 
-    const { verboseLog, startUrls = [] } = input;
+    const {
+        verboseLog,
+        startUrls = [],
+        proxyConfiguration,
+        searchKeywords,
+        maxResults,
+        postsFromDate,
+        handlePageTimeoutSecs = 3600,
+    } = input;
     if (verboseLog) {
         log.setLevel(log.LEVELS.DEBUG);
     }
 
-    // launch options - puppeteer
-    const pptrLaunchOpts = {};
-    pptrLaunchOpts.stealth = false; // TODO: change back when this is fixed, disable for increased performance
-    pptrLaunchOpts.useChrome = Apify.isAtHome();
-
     const requestQueue = await Apify.openRequestQueue();
+    const proxyConfig = await utils.proxyConfiguration({
+        proxyConfig: proxyConfiguration,
+    });
 
-    /**
-     * @type {Apify.PuppeteerCrawlerOptions}
-     */
-    const pptrCrawlerOpts = {};
-    pptrCrawlerOpts.requestQueue = requestQueue;
-    pptrCrawlerOpts.launchPuppeteerOptions = pptrLaunchOpts;
-    pptrCrawlerOpts.useSessionPool = true;
-    pptrCrawlerOpts.proxyConfiguration = await Apify.createProxyConfiguration({ ...input.proxyConfiguration });
-
-    pptrCrawlerOpts.launchPuppeteerFunction = crawler.hndlPptLnch;
-    pptrCrawlerOpts.gotoFunction = crawler.hndlPptGoto;
-    pptrCrawlerOpts.handlePageTimeoutSecs = 3600;
-    pptrCrawlerOpts.handleFailedRequestFunction = crawler.hndlFaildReqs;
-    pptrCrawlerOpts.handlePageFunction = async ({ page, request, puppeteerPool, response }) => {
-        const hasCaptcha = await page.$('.g-recaptcha');
-        if (hasCaptcha) {
-            await puppeteerPool.retire(page.browser());
-            throw 'Got captcha, page will be retried. If this happens often, consider increasing number of proxies';
-        }
-
-        if (utils.isErrorStatusCode(response.status())) {
-            await puppeteerPool.retire(page.browser());
-            throw `Response status is: ${response.status()} msg: ${response.statusText()}`;
-        }
-
-        switch (request.userData.label) {
-            case 'CHANNEL':
-            case 'SEARCH':
-            case 'MASTER': {
-                await crawler.handleMaster(page, requestQueue, input, request);
-                break;
-            }
-            case 'DETAIL': {
-                await crawler.handleDetail(page, request);
-                break;
-            }
-            default: throw new Error('Unknown request label in handlePageFunction');
-        }
-    };
-
-    if (!input.searchKeywords && (!startUrls || !startUrls.length)) {
+    if (!searchKeywords && (!startUrls || !startUrls.length)) {
         throw new Error('You need to provide either searchKeywords or startUrls as input');
     }
 
@@ -99,19 +68,130 @@ Apify.main(async () => {
                 },
             });
         }
-    } else if (input.searchKeywords) {
+    } else if (searchKeywords) {
         // add starting url
         log.info('Starting scraper with a search keyword');
 
-        await requestQueue.addRequest({
-            url: 'https://www.youtube.com/',
-            userData: {
-                label: 'MASTER',
-                search: input.searchKeywords,
-            },
-        });
+        for (let searchKeyword of searchKeywords.split(',')) {
+            searchKeyword = `${searchKeyword}`.trim();
+
+            if (searchKeyword) {
+                await requestQueue.addRequest({
+                    url: 'https://www.youtube.com/',
+                    uniqueKey: `SEARCH-${searchKeyword}`,
+                    userData: {
+                        label: 'MASTER',
+                        search: searchKeyword.trim(),
+                    },
+                });
+            }
+        }
     }
 
-    const pptrCrawler = new Apify.PuppeteerCrawler(pptrCrawlerOpts);
+    const extendOutputFunction = await utils.extendFunction({
+        input,
+        key: 'extendOutputFunction',
+        output: async (data) => {
+            await Apify.pushData(data);
+        },
+        helpers: {},
+    });
+
+    const extendScraperFunction = await utils.extendFunction({
+        input,
+        key: 'extendScraperFunction',
+        output: async () => {}, // no-op for page interaction
+        helpers: {
+            requestQueue,
+            extendOutputFunction,
+        },
+    });
+
+    const pptrCrawler = new Apify.PuppeteerCrawler({
+        requestQueue,
+        browserPoolOptions: {
+            maxOpenPagesPerBrowser: 1,
+        },
+        useSessionPool: true,
+        proxyConfiguration: proxyConfig,
+        preNavigationHooks: [
+            async ({ page }) => {
+                await puppeteer.blockRequests(page, {
+                    urlPatterns: [
+                        '.mp4',
+                        '.webp',
+                        '.jpeg',
+                        '.jpg',
+                        '.gif',
+                        '.svg',
+                        '.ico',
+                        'google-analytics',
+                        'doubleclick.net',
+                        'googletagmanager',
+                        '/videoplayback',
+                        '/adview',
+                        '/stats/ads',
+                        '/stats/watchtime',
+                        '/stats/qoe',
+                        '/log_event',
+                    ],
+                });
+            },
+        ],
+        handlePageTimeoutSecs,
+        handleFailedRequestFunction: async ({ request }) => {
+            Apify.utils.log.error(`Request ${request.url} failed too many times`);
+
+            await Apify.pushData({
+                '#debug': Apify.utils.createRequestDebugInfo(request),
+            });
+        },
+        handlePageFunction: async ({ page, request, session, response }) => {
+            // no-output function
+            await extendScraperFunction(undefined, {
+                page,
+                request,
+            });
+
+            const hasCaptcha = await page.$('.g-recaptcha');
+            if (hasCaptcha) {
+                session.retire();
+                throw 'Got captcha, page will be retried. If this happens often, consider increasing number of proxies';
+            }
+
+            if (utils.isErrorStatusCode(response.status())) {
+                session.retire();
+                throw `Response status is: ${response.status()} msg: ${response.statusText()}`;
+            }
+
+            if (await page.$('.yt-upsell-dialog-renderer')) {
+                // this dialog steal focus, so need to click it
+                await page.evaluate(async () => {
+                    const noThanks = document.querySelectorAll('.yt-upsell-dialog-renderer [role="button"]');
+
+                    for (const button of noThanks) {
+                        if (button.textContent && button.textContent.includes('No thanks')) {
+                            button.click();
+                            break;
+                        }
+                    }
+                });
+            }
+
+            switch (request.userData.label) {
+                case 'CHANNEL':
+                case 'SEARCH':
+                case 'MASTER': {
+                    await crawler.handleMaster({ page, requestQueue, searchKeywords, maxResults, postsFromDate, request });
+                    break;
+                }
+                case 'DETAIL': {
+                    await crawler.handleDetail(page, request, extendOutputFunction);
+                    break;
+                }
+                default: throw new Error('Unknown request label in handlePageFunction');
+            }
+        },
+    });
     await pptrCrawler.run();
 });