+++ src/utility.js
@@ -1,13 +1,14 @@
 const moment = require('moment');
 const Apify = require('apify');
 const Puppeteer = require('puppeteer'); // eslint-disable-line
+const vm = require('vm');
 
 const { log, sleep } = Apify.utils;
 
 const CONSTS = require('./consts');
 
 exports.handleErrorAndScreenshot = async (page, e, errorName) => {
-    await Apify.utils.puppeteer.saveSnapshot(page, { key: `ERROR-${errorName}-${Math.random()}`});
+    await Apify.utils.puppeteer.saveSnapshot(page, { key: `ERROR-${errorName}-${Math.random()}` });
     throw `Error: ${errorName} - Raw error: ${e.message}`;
 };
 
@@ -27,7 +28,7 @@ exports.loadVideosUrls = async (requestQueue, page, inputUrl, maxRequested, isSe
 
     const logInterval = setInterval(
         () => log.info(`[${searchOrUrl}]: Scrolling state - Enqueued ${videosEnqueuedUnique} unique video URLs, ${videosEnqueued} total`),
-        60000
+        60000,
     );
 
     try {
@@ -45,7 +46,9 @@ exports.loadVideosUrls = async (requestQueue, page, inputUrl, maxRequested, isSe
                 }
 
                 for (const video of videos) {
-                    await video.hover();
+                    try {
+                        await video.hover();
+                    } catch (e) {}
 
                     const rq = await requestQueue.addRequest({
                         url: await video.$eval(url, (el) => el.href),
@@ -100,7 +103,7 @@ exports.getDataFromXpath = async (page, xPath, attrib) => {
 };
 
 exports.getDataFromSelector = async (page, slctr, attrib) => {
-    const slctrElem = await page.waitForSelector(slctr, { visible: true, timeout: 120000 });
+    const slctrElem = await page.waitForSelector(slctr, { visible: true, timeout: 60000 });
     return page.evaluate((el, key) => el[key], slctrElem, attrib);
 };
 
@@ -200,7 +203,7 @@ exports.clickHoveredElem = async (pptPage, xPath) => {
  */
 exports.doTextInput = async (pptPage, keywords) => {
     for (let i = 0; i < keywords.length; i++) {
-        await pptPage.type('input#search', keywords[i], { delay: CONSTS.DELAY.BTWN_KEY_PRESS });
+        await pptPage.type('input#search', keywords[i], { delay: CONSTS.DELAY.BTWN_KEY_PRESS.max });
         await Apify.utils.sleep(exports.getDelayMs(CONSTS.DELAY.BTWN_KEY_PRESS));
     }
 };
@@ -218,15 +221,21 @@ exports.getCutoffDate = (historyString) => {
     return moment().subtract(numDurations, durationType);
 };
 
+/**
+ * @param {string} postsFromDate
+ */
 exports.isDateInputValid = (postsFromDate) => {
     if (postsFromDate) {
-        const matches = postsFromDate.match(/(^(1|([^0a-zA-Z ][0-9]{0,3})) (minute|hour|day|week|month|year))s? ago *$/ig);
+        const matches = postsFromDate.match(/(^(1|([^0a-zA-Z ][0-9]{0,3})) (minute|hour|day|week|month|year))s?/ig);
         return !!matches;
     }
 
     return false;
 };
 
+/**
+ * @param {string} postsFromDate
+ */
 exports.getYoutubeDateFilters = (postsFromDate) => {
     if (!exports.isDateInputValid(postsFromDate)) {
         return [];
@@ -263,7 +272,7 @@ exports.getYoutubeDateFilters = (postsFromDate) => {
     if (youtubeFilters.length > 0) {
         // if we are using any of the youtube date filters
         // then we must also sort results by 'Upload date'
-        youtubeFilters.push('Upload date');
+        youtubeFilters.unshift('Upload date');
     }
 
     return youtubeFilters;
@@ -330,4 +339,165 @@ exports.getRandBetween = (min, max) => {
 
 exports.getDelayMs = (minMax) => {
     return exports.getRandBetween(minMax.min, minMax.max);
+};
+
+/**
+ * @template T
+ * @typedef {T & { Apify: Apify, customData: any, request: Apify.Request }} PARAMS
+ */
+
+/**
+ * Compile a IO function for mapping, filtering and outputing items.
+ * Can be used as a no-op for interaction-only (void) functions on `output`.
+ * Data can be mapped and filtered twice.
+ *
+ * Provided base map and filter functions is for preparing the object for the
+ * actual extend function, it will receive both objects, `data` as the "raw" one
+ * and "item" as the processed one.
+ *
+ * Always return a passthrough function if no outputFunction provided on the
+ * selected key.
+ *
+ * @template RAW
+ * @template {{ [key: string]: any }} INPUT
+ * @template MAPPED
+ * @template {{ [key: string]: any }} HELPERS
+ * @param {{
+ *  key: string,
+ *  map?: (data: RAW, params: PARAMS<HELPERS>) => Promise<MAPPED>,
+ *  output?: (data: MAPPED, params: PARAMS<HELPERS> & { data: RAW, item: MAPPED }) => Promise<void>,
+ *  filter?: (obj: { data: RAW, item: MAPPED }, params: PARAMS<HELPERS>) => Promise<boolean>,
+ *  input: INPUT,
+ *  helpers: HELPERS,
+ * }} params
+ * @return {Promise<(data: RAW, args?: Record<string, any>) => Promise<void>>}
+ */
+const extendFunction = async ({
+    key,
+    output,
+    filter,
+    map,
+    input,
+    helpers,
+}) => {
+    /**
+     * @type {PARAMS<HELPERS>}
+     */
+    const base = {
+        ...helpers,
+        Apify,
+        customData: input.customData || {},
+    };
+
+    const evaledFn = (() => {
+        // need to keep the same signature for no-op
+        if (typeof input[key] !== 'string' || input[key].trim() === '') {
+            return new vm.Script('({ item }) => item');
+        }
+
+        try {
+            return new vm.Script(input[key], {
+                lineOffset: 0,
+                produceCachedData: false,
+                displayErrors: true,
+                filename: `${key}.js`,
+            });
+        } catch (e) {
+            throw new Error(`"${key}" parameter must be a function`);
+        }
+    })();
+
+    /**
+     * Returning arrays from wrapper function split them accordingly.
+     * Normalize to an array output, even for 1 item.
+     *
+     * @param {any} value
+     * @param {any} [args]
+     */
+    const splitMap = async (value, args) => {
+        const mapped = map ? await map(value, args) : value;
+
+        if (!Array.isArray(mapped)) {
+            return [mapped];
+        }
+
+        return mapped;
+    };
+
+    return async (data, args) => {
+        const merged = { ...base, ...args };
+
+        for (const item of await splitMap(data, merged)) {
+            if (filter && !(await filter({ data, item }, merged))) {
+                continue; // eslint-disable-line no-continue
+            }
+
+            const result = await (evaledFn.runInThisContext()({
+                ...merged,
+                data,
+                item,
+            }));
+
+            for (const out of (Array.isArray(result) ? result : [result])) {
+                if (output) {
+                    if (out !== null) {
+                        await output(out, { ...merged, data, item });
+                    }
+                    // skip output
+                }
+            }
+        }
+    };
+};
+
+exports.extendFunction = extendFunction;
+
+/**
+ * Do a generic check when using Apify Proxy
+ *
+ * @typedef params
+ * @property {any} [params.proxyConfig] Provided apify proxy configuration
+ * @property {boolean} [params.required] Make the proxy usage required when running on the platform
+ * @property {string[]} [params.blacklist] Blacklist of proxy groups, by default it's ['GOOGLE_SERP']
+ * @property {boolean} [params.force] By default, it only do the checks on the platform. Force checking regardless where it's running
+ * @property {string[]} [params.hint] Hint specific proxy groups that should be used, like SHADER or RESIDENTIAL
+ *
+ * @param {params} params
+ * @returns {Promise<Apify.ProxyConfiguration | undefined>}
+ */
+module.exports.proxyConfiguration = async ({
+    proxyConfig,
+    required = true,
+    force = Apify.isAtHome(),
+    blacklist = ['GOOGLESERP'],
+    hint = [],
+}) => {
+    const configuration = await Apify.createProxyConfiguration(proxyConfig);
+
+    // this works for custom proxyUrls
+    if (Apify.isAtHome() && required) {
+        if (!configuration
+            || (!configuration.usesApifyProxy
+                && (!configuration.proxyUrls || !configuration.proxyUrls.length)) || !configuration.newUrl()) {
+            throw '\n=======\nWrong Input! You must use Apify proxy or custom proxies with this scraper!\n\n=======';
+        }
+    }
+
+    // check when running on the platform by default
+    if (force) {
+        // only when actually using Apify proxy it needs to be checked for the groups
+        if (configuration && configuration.usesApifyProxy) {
+            if (blacklist.some((blacklisted) => (configuration.groups || []).includes(blacklisted))) {
+                throw '\n=======\nThese proxy groups cannot be used in this actor.'
+                    + `Choose other group or contact support@apify.com to give you proxy trial:\n\n*  ${blacklist.join('\n*  ')}\n\n=======`;
+            }
+
+            // specific non-automatic proxy groups like RESIDENTIAL, not an error, just a hint
+            if (hint.length && !hint.some((group) => (configuration.groups || []).includes(group))) {
+                Apify.utils.log.info(`\n=======\nYou can pick specific proxy groups for better experience:\n\n*  ${hint.join('\n*  ')}\n\n=======`);
+            }
+        }
+    }
+
+    return configuration;
 };